Add User, Product logic

+++ src/main/java/dev/prom/delivery/service/UserServiceImpl.java
@@ -1,12 +1,13 @@
 package dev.prom.delivery.service;
 
-import dev.prom.delivery.dto.OrderInputDto;
-import dev.prom.delivery.dto.UserDto;
+import dev.prom.delivery.exceptions.NotFoundException;
 import dev.prom.delivery.models.User;
 import dev.prom.delivery.repository.UserRepository;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
@@ -17,27 +18,31 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else throw new NotFoundException("User not found");
     }
 
     @Override
     public User updateUser(Long id, User user) {
-        user.setId(id);
-        return userRepository.save(user);
+        if (userRepository.existsById(id)) {
+            user.setId(id);
+        } else throw new NotFoundException("User not found");
+        return userRepository.saveAndFlush(user);
     }
 
     @Override
     public User createUser(User user) {
-        return null;
+        return userRepository.saveAndFlush(user);
     }
 
     @Override
     public User getUserById(Long id) {
-        return null;
+        return userRepository.findById(id).orElseThrow(() -> new NotFoundException("User not found"));
     }
 
     @Override
     public List<User> getAllUsers() {
-        return null;
+        return userRepository.findAll();
     }
 }
